remove unneccessary casts

+++ src/freenet/client/async/SplitFileFetcherSegment.java
@@ -289,7 +289,7 @@ public class SplitFileFetcherSegment implements FECCallback, HasCooldownTrackerI
 					copy = Long.MAX_VALUE;
 				else
 					copy = truncateLength - totalCopied;
-				if(copy < ((long)CHKBlock.DATA_LENGTH))
+				if(copy < (CHKBlock.DATA_LENGTH))
 					buf = new byte[(int)copy];
 				InputStream is = data.getInputStream();
 				DataInputStream dis = new DataInputStream(is);
