+++ src/freenet/support/compress/NewLZMACompressor.java
@@ -114,7 +114,7 @@ public class NewLZMACompressor implements Compressor {
 		
 		int dictionarySize = 0;
 		for (int i = 0; i < 4; i++)
-			dictionarySize += ((int)(props[1 + i]) & 0xFF) << (i * 8);
+			dictionarySize += ((props[1 + i]) & 0xFF) << (i * 8);
 		
 		if(dictionarySize < 0) throw new InvalidCompressedDataException("Invalid dictionary size");
 		if(dictionarySize > MAX_DICTIONARY_SIZE) throw new TooBigDictionaryException();
